feat: :rocket: check syntax errors and refactor imported code
There is the file buggy.py which have some code that needs to be refactored. There is code_refactor.py that searches for non-closed strings or brackets. It also refactors code in a way that makes it easy to test it or turn it to a graph. Read comments for more.