fix: :bug: Fixed bugs in CFG functions "extractAllDefs" and "construct_graph"
extractAllDefs: was not working properly when a function is defined right after another function definition ends. construct_graph:  was not correctly identifying elifs and elses, so points the last line of the if block to the elif or else statements even though it should skip the entire elif or else block.