refactor: :recycle: Modified extractAllDefs()
It now only extracts function definitions that can be accessed globally by the code in the graph (let's call it current), makes a graph of it (let's call it child1) and adds it to the current graph children list. When child1 calls extractAllDefs(), it will do the same thing for it and add the new graph (child1.1) to this child1 children list, and so on.